# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: AI Agent Platform API
  description: Enterprise AI Agent Platform API Documentation
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@aiagentplatform.com
  license:
    name: Proprietary
    url: https://aiagentplatform.com/license
servers:
  - url: https://api.aiagentplatform.com/v1
    description: Production server
  - url: https://staging-api.aiagentplatform.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    CreateCheckoutRequest:
      type: object
      properties:
        planId:
          type: string
          enum: [STARTER, PROFESSIONAL, ENTERPRISE]
        interval:
          type: string
          enum: [month, year]
        successUrl:
          type: string
          format: uri
        cancelUrl:
          type: string
          format: uri
        couponCode:
          type: string
      required:
        - planId
        - interval
    CreateCheckoutResponse:
      type: object
      properties:
        sessionId:
          type: string
        sessionUrl:
          type: string
          format: uri
paths:
...
  /billing/create-checkout:
    post:
      tags: [Billing]
      summary: Create Stripe checkout session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCheckoutRequest'
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCheckoutResponse'
  /billing/subscription:
    get:
      tags: [Billing]
      summary: Get current subscription
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subscription details

    put:
      tags: [Billing]
      summary: Update subscription
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                planId:
                  type: string
                interval:
                  type: string
      responses:
        '200':
          description: Subscription updated

    delete:
      tags: [Billing]
      summary: Cancel subscription
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subscription cancelled

  # Admin
...